requirements.txt
streamlit
python-dotenv
elevenlabs
openai